[SOFT-46] Move event_id to blink_event_generator_start and improve semantics
The API and semantics now more closely align with lights_blinker from MSXIII.

event_id is now passed to blink_event_generator_start and isn't a field in
BlinkEventGeneratorSettings, which is better for object reusability. We also
automatically stop blinking if blink_event_generator_start is called before
blink_event_generator_stop and a blinker is running, and on every
blink_event_generator_stop we default to the first_value specified in the settings.