[SOFT-46] Improve blink_event_generator behaviour
When blink_event_generator_stop is called and the blinker isn't in the
default state, raise a final event to transition back to the default state.

When blink_event_generator_start is called, the blinker is running, and
the event_id is equal to the blinker's event_id (so we aren't changing
events), do nothing to avoid a delay in the blinking.