[SOFT-240] Implement deterministic soft timers
Running tests takes valuable developer time, which we want
to minimize. To reduce it, we can mock the passing of time
with counters rather than actual timers.

This commit implements these counters as well as adds them
to a sample test, adc_periodic_reader, to demonstrate
functionality. Also made a slight fix to the test since
it was previously broken by the x86 implementation of adc.c.