[SOFT-171] Implement BMS relay controls
This implements BMS relay controls as sequences rather than an FSM, since things need to happen
asynchronously, an FSM would over-complicate things.
To close relays, the steps are:
1. close ground
2. assert it closed properly, if not fault
3. close hv
4. assert it closer properly, if not fault
5. broadcast done.
A similar procedure takes place for opening, but we open them at the same time.
We need to stagger the closing of the relays to escape simultaneous inrush current, which
could trigger overcurrent protection.