[SOFT-336] Horrible hack for quick PD BTS7200/7040 update
The BTS7200/7040 ("BTS7xxx") drivers were updated recently such that for
the fault handling logic to work correctly, all manipulation of the load
switch enable pins must be done through the driver. I updated power
distro to conform with a horrible hack involving retrieving an array of
Bts7xxxEnablePin structs from current_measurement directly from pd_gpio
and linear-searching through it to find a pin to set...eventually this
will be replaced with an actual good solution, but in the meantime this
will work for the mockup.

This is on a separate branch so it can be reverted should everything
break, and so it's easier to merge into master (eventually).