[SOFT-333] Account for hazard button latching
The hazard button on centre console latches, so this updates button_press
and hazard_tx to account for this. Now hazard button press events (as well as all
button press events) have the GpioState of the button pin as the event data.
hazard_tx uses the state of the hazard button pin instead of an internal bool to
keep decide whether hazards are on or not.

I assumed that the hazard button is active-high; if it's active-low, change
HAZARD_BUTTON_ACTIVE_STATE in hazard_tx.h to GPIO_STATE_LOW.