[SOFT-349] Store MPPT status instead of just CR bit
The data store previously stored just the mysterious CR bit, which is
the last bit of the MPPT status. For fan control we need other parts of
the status. It's easiest to just store the whole thing and extract parts
as needed with the spv1020_is_* functions from the SPV1020 driver. So
now we store the whole status in the data store instead of just the CR
bit.