Make test_data_tx faster and less flaky
Due to the CAN bug, data_tx's tests previously had to tx only one
message at a time, with 350ms between each tx. This was *super* slow,
and because there were ~50 messages to tx, super flaky. This changes the
test to tx 8 at a time, 150ms apart, which is way faster (<1s instead of
10+ seconds) and less flaky.