[SOFT-301] Fix race condition in x86 CAN library (#290)
This fixes the inconsistently failing CAN unit tests and the majority of the instability of CAN on x86.

The culprit seems to be prv_tx_handler in can.c. It’s a callback which runs every time the can_hw implementation registers a CAN_HW_EVENT_TX_READY (which is just a condition where the callback is called, not an event queue event). On stm32, there are 3 CAN mailboxes, so on each TX event can_hw tries to TX, and if there are no mailboxes it swallows the event (but does not pop from the CAN fifo) and exits. The prv_tx_handler callback reraises the TX event on the next TX when this happens (which is when the CAN fifo is nonempty and a TX occurs) so that TXes aren’t lost.

However, on x86, there are no CAN mailboxes and we have a TX thread that’s consuming TXes (from the TX fifo, which isn’t the same as the CAN fifo) and putting them on the vcan bus. Whenever the TX thread completes a TX, it of course calls prv_tx_handler as the TX callback. Meanwhile, whenever a TX event gets processed, can_hw_transmit and prv_handle_tx in can_fsm pop a message from the CAN fifo, feed it to the TX fifo, and signal the TX thread with a semaphore. Sometimes, the TX thread takes its sweet time waking up when the semaphore is raised, and sometimes it wakes up right away.

This creates a race condition. If all the TX events can be processed and the CAN fifo emptied before the TX thread wakes up for the first time, then the CAN fifo is empty when prv_tx_handler is called first, and no extra TX event is raised. However, if the TX thread wakes up (and completes its ~250us sleep after the first write) before all the TX events are processed, then the CAN fifo is nonempty for prv_tx_handler and an extraneous TX event is raised, which a) messes up tests using MS_TEST_HELPER_CAN_TX_RX in a big way and b) even if it doesn’t, the TX that the extra event is supposed to save will already have been TX’d by the time the extra event gets processed, and tons of weird things could probably happen. This is what causes all the test instability, like when I added a LOG_DEBUG to a MS_TEST_HELPER_CAN_TX loop and suddenly everything started failing.

Fixed by adding an X86 macro defined on x86 only and using it to make prv_tx_handler only do its thing on stm32.

* Fix PD's publish_data for CAN changes

publish_data was sending 22 CAN messages at once, which exhausted the
event queue (size 20) but not the CAN tx fifo (size 32), so previously
the bug in can.c was saving it. Change it to send them incrementally
like solar/data_tx, now it works.

* Make test_data_tx faster and less flaky

Due to the CAN bug, data_tx's tests previously had to tx only one
message at a time, with 350ms between each tx. This was *super* slow,
and because there were ~50 messages to tx, super flaky. This changes the
test to tx 8 at a time, 150ms apart, which is way faster (<1s instead of
10+ seconds) and less flaky.

Co-authored-by: mitchellostler <ostlermitchell44@gmail.com>