[SOFT-416] Minimal bootloader with linker scripts (#273)
This adds a minimal bootstrapped version of the bootloader. The bootloader itself does nothing except jump to the application code. The bulk of the complexity is in the linker scripts for organizing the bootloader and application code in memory.

There's also a provision for selecting stm32 linker scripts per project from rules.mk (mostly useful for the bootloader project), and a temporary makefile target, temp-bootloader-write, for building the bootloader and application code and flashing both together.