[SOFT-447] Patch x86 CAN library to allow a message with a zero CAN ID (#318)
On x86, the flow for receiving a CAN message is the following:
  1. The RX thread (prv_rx_thread in can_hw.c) reads a CAN frame from the socket.
  2. The RX thread calls a handler, which turns out to be prv_rx_handler in can.c.
  3. prv_rx_handler calls can_hw_receive until it returns false. On x86, can_hw_receive will only rx one message (the most recently rx'd one), but it does this to support STM32 which can receive multiple in a row.
  4. On the first call, can_hw_receive sees that it has a valid frame received, passes the data upwards, resets it, and returns true; on the second call, it sees that it doesn't have a valid frame and returns false.

Previously, we determined whether the frame was valid by assuming that a CAN arbitration ID of 0 would never happen, so the zero ID caused by memsetting the frame to 0 was checked. Now with the bootloader, an ID of 0 is possible (and likely), so we use a separate flag to determine if the frame is valid.

Note that `read` returns -1 when it fails to read; we don't want to process a frame when it wasn't read correctly, so we check read's return value when setting the is-valid flag.