Rewrite test_can_wake_works to avoid infinite looping
There was a very odd race condition in test_can_wake_works where very
occasionally it would loop forever. Not sure why, but it also wasn't
testing what it was supposed to particularly effectively, so I rewrote
it.

Also added a couple of x86_interrupt_wake() calls in can_hw to be safe.