Initialize, Register, & Implement CAN RX Handler
fault_handler doesn’t actually use the event queue, it TXes over CAN (which is what you want, to tell telemetry that there’s a fault). In particular it calls either CAN_TRANSMIT_SOLAR_FAULT_5_MPPTS or CAN_TRANSMIT_SOLAR_FAULT_6_MPPTS depending on whether you initialize it as 5 or 6 mppts. Now what a CAN_TRANSMIT_* macro does is it raises a TX event, which has to be processed; then when it receives the event back (you’ll set up CAN in loopback mode so it does this), it raises an RX event, which has to be processed, and then any handlers you register for those CAN messages will be called. - Coppernium