[SOFT-349][SOFT-425] Fan control for solar (#327)
* [SOFT-349] Store MPPT status instead of just CR bit

The data store previously stored just the mysterious CR bit, which is
the last bit of the MPPT status. For fan control we need other parts of
the status. It's easiest to just store the whole thing and extract parts
as needed with the spv1020_is_* functions from the SPV1020 driver. So
now we store the whole status in the data store instead of just the CR
bit.

* [SOFT-349] Fan control for solar

This adds the fan_control_solar module to solar, which does two things
to control the fans on solar:
- raise a fault whenever the fan fails or goes overtemperature
- turn the fans to full speed whenever an mppt or a thermistor reports
  overtemp

Co-authored-by: Ryan Dancy <ryan@keal.ca>
Co-authored-by: Patrick Kim <s524kim@uwaterloo.ca>
Co-authored-by: FourRhythm <FourRhythm@users.noreply.github.com>
Co-authored-by: Ryan Dancy <41841896+copperium@users.noreply.github.com>