[SOFT-455] Implement chip id (#349)
STM32s have a factory-set 96-bit unique chip ID in a system register. See section 33.1 of the STM32F0xx manual.

This could be useful in some contexts (e.g. maybe the bootloader), so this adds an ms-common module which retrieves the chip ID and exposes it as a custom data type.

Co-authored-by: Jackie Zhu <64090163+jckpokimain@users.noreply.github.com>