[SOFT-456] Brake Light Control
In centre console, pedal_monitor.c receives updates from pedal and decides whether the pedal is pressed or released. Use this to tell power distribution when to turn the brake light on and off.

To get PD to turn the brake light off, transmit a SYSTEM_CAN_MESSAGE_LIGHTS with the lights ID (first u8) as EE_LIGHTS_TYPE_BRAKES and the state (second u8) as either EE_LIGHTS_STATE_OFF or EE_LIGHTS_STATE_ON.

To save bandwidth on the CAN network, only transmit this when the state changes.