[SOFT-380] Fix buffer overflow in MU logging (#360)
The MU LOG implementation was using a static buffer to store logs to be exported. This was causing issues in PD, where use of the BUG macro would cause strings to exceed the 256-byte buffer (and gcc would actually detect this at compile time!).

To fix this, we get rid of the buffer entirely by using asprintf, a GNU extension which works like snprintf except it mallocs enough memory for the resulting string automatically. This does mean using the heap, but that's ok on MU since it's only for x86. Plus, we now don't have a `static char s_log_buf[256]` polluting the namespace of every file that includes log.h on MU.