[SOFT-461] Assert when solar relay opens (#399)
When the solar relay gets opened, we wait an amount of time then assert that it really was opened. We declare a failure and send a `EE_SOLAR_RELAY_OPEN_ERROR` fault if the current through the array is greater than a threshold, there was a previous DRV120 relay error, or the current isn't set after a while. Otherwise, we send a "ready to drive" message.