[SOFT-275] MCI Fan Control (#419)
Fan control module for MCI. The way this works is pretty simple -- there's a single pin, MCI_FAN_EN_ADDR, that is used to turn on/off the fans. This is abstracted by mci_fan_set_state(). Hardware wants the fan state based on the temperature measurements from the motor controllers, so that's for a future ticket.

There are also four pins that go high when the corresponding thermistor on the MCI board hits a certain threshold. Eventually we should log these over CAN using SYSTEM_CAN_MESSAGE_MOTOR_STATUS, but for now we just track their status in a fault bitset that can be exported using mci_fan_get_fault_bitset(). A fault callback can also be passed in on initialization, and will be called each time the fault bitset updates.

The test sequence for this is pretty short, as the module itself is quite simple.