[SOFT-504] Change front PD adc read scale (#420)
For front PD, we've currently got a linear scale with 0 to the ADC reference voltage mapping to 0-100% fan speed. Per Aashmika and onsite testing, we need 0-1.65V to map to 0-100% fan speed. (1.65V or 1650mV should be a configurable parameter in a #define at the top of the file.) Also,
> if it is 3.3V the fans should be 0% or off (because the spst (switch) is open)
But the ADC can only detect up to its reference voltage, which is ~2V or so usually. So instead say that if the voltage is >1.65V (by some epsilon, maybe 50mV) and within an epsilon (again maybe 50mV) of min(reference voltage, 3.3V) then say that the switch is open and shut the fans off.