[SOFT-515] Add custom BTS7004 scaling for UV VBAT (#436)
PD reads currents from the UV VBAT pin from UV cutoff. It's on a BTS7004 load switch, which is very similar to a BTS7040 load switch but uses different scaling factors. This adds a BTS7004 scaling factor and a flag to switch between them. The scaling factor will have to be calibrated onsite, currently it's a theoretical value from the datasheet.